requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
cssselect2==0.8.0
defusedxml==0.7.1
Django==4.2.23
django-axes==8.0.0
django-ipware==4.0.2
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
django-widget-tweaks==1.5.0
fonttools==4.27.1
html5lib==1.1
idna==3.10
lxml==6.0.1
oscrypto==1.3.0
pillow==11.3.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
PyMySQL==1.1.1
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
six==1.16.0
sqlparse==0.4.4
svg.path==6.2
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.1.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
xhtml2pdf==0.2.17
zopfli==0.2.3.post1